app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.0.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
